backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
python-dotenv==1.0.0
bcrypt==4.0.1
Faker==19.6.2
pytest==7.4.2
pytest-flask==1.2.0
requests==2.31.0
google-generativeai==0.8.3
